move include/common to include/finufft_common
