--- include/common/common.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <common/constants.h>
-#include <common/defines.h>
-#include <common/utils.h>
