+++ include/cufinufft/utils.h
@@ -12,7 +12,7 @@
 #include <type_traits>
 #include <utility> // for std::forward
 
-#include <common/common.h>
+#include <finufft_common/common.h>
 
 #ifndef _USE_MATH_DEFINES
 #define _USE_MATH_DEFINES
@@ -137,7 +137,7 @@ template<typename T> auto arrayrange(int n, T *a, cudaStream_t stream) {
 
 // Writes out w = half-width and c = center of an interval enclosing all a[n]'s
 // Only chooses a nonzero center if this increases w by less than fraction
-// ARRAYWIDCEN_GROWFRAC defined in common/constants.h.
+// ARRAYWIDCEN_GROWFRAC defined in finufft_common/constants.h.
 // This prevents rephasings which don't grow nf by much. 6/8/17
 // If n==0, w and c are not finite.
 template<typename T> auto arraywidcen(int n, T *a, cudaStream_t stream) {
