+++ include/finufft/finufft_utils.hpp
@@ -5,7 +5,7 @@
 
 #include "finufft_core.h"
 #include <cmath>
-#include <common/common.h>
+#include <finufft_common/common.h>
 
 namespace finufft::utils {
 
@@ -28,7 +28,7 @@ FINUFFT_EXPORT FINUFFT_ALWAYS_INLINE void FINUFFT_CDECL arraywidcen(BIGINT n, co
                                                                     T *w, T *c)
 // Writes out w = half-width and c = center of an interval enclosing all a[n]'s
 // Only chooses a nonzero center if this increases w by less than fraction
-// ARRAYWIDCEN_GROWFRAC defined in common/constants.h.
+// ARRAYWIDCEN_GROWFRAC defined in finufft_common/constants.h.
 // This prevents rephasings which don't grow nf by much. 6/8/17
 // If n==0, w and c are not finite.
 {
