+++ include/finufft_common/common.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <finufft_common/constants.h>
+#include <finufft_common/defines.h>
+#include <finufft_common/utils.h>
