+++ test/cuda/cufinufft1dspreadinterponly_test.cu
@@ -4,7 +4,7 @@
 #include <iostream>
 #include <random>
 
-#include <common/common.h>
+#include <finufft_common/common.h>
 #include <cufinufft.h>
 
 #include <cufinufft/contrib/helper_cuda.h>
