+++ test/cuda/cufinufft2dmany_test.cu
@@ -6,7 +6,7 @@
 #include <limits>
 #include <random>
 
-#include <common/common.h>
+#include <finufft_common/common.h>
 #include <cufinufft.h>
 
 #include <cufinufft/impl.h>
